+++ tokens/internal/varsig/varsig.go
